+++ web/src/components/planner/PlannerSidebarContent.vue
@@ -13,12 +13,10 @@
   <v-divider color="#ccc" thickness="2px" />
   <planner-global-actions
     class="py-2"
-    :help-text-shown="helpTextShown"
     @clear-all="emit('clearAll')"
     @hide-all="emit('hideAll')"
     @import-world="emit('importWorld')"
     @show-all="emit('showAll')"
-    @toggle-help-text="emit('toggleHelpText')"
   />
   <v-divider color="#ccc" thickness="2px" />
   <copyright />
@@ -31,7 +29,6 @@
   defineProps<{
     factories: Factory[],
     loadedFrom: 'planner' | 'navigation',
-    helpTextShown: boolean,
   }>()
 
   const emit = defineEmits<{
@@ -41,6 +38,5 @@
     (event: 'hideAll'): void;
     (event: 'showAll'): void;
     (event: 'importWorld'): void;
-    (event: 'toggleHelpText'): void;
   }>()
 </script>