+++ web/src/components/planner/products/ProductsAndPower.vue
@@ -3,19 +3,10 @@
     <div class="d-flex align-center flex-wrap mb-4 ga-2">
       <h1 class="text-h5" :class="heading.class">
         <i :class="heading.icon" />
-        <span class="ml-3">Products &amp; Power Generators</span>
+        <span class="ml-3">Products, Power &amp; Buildings</span>
       </h1>
     </div>
-    <p v-show="helpText" class="text-body-2 mb-4">
-      <i class="fas fa-info-circle" /> Products that are created within the factory. Products are first
-      used to fulfil recipes internally, and any surplus is then available for Export.<br>
-      e.g. if you add 200 Iron Rods and also 100 Screws, you'd have 100 surplus Rods remaining used as an
-      Export (and the Screws as a end product).<br>
-      An <v-chip color="green">Internal</v-chip> product is one that is used to produce other products. The surplus of which can also be used as an export.<br>
-      A <v-chip class="sf-chip status-note"><i class="fas fa-question-circle mr-1" />No demand</v-chip> product is one nothing asks for: not used internally, not exported. A future update will add support for sinking, so if you are sinking it, ignore this for now.<br>
-      A <v-chip class="sf-chip status-warning"><i class="fas fa-exclamation-triangle mr-1" />Potential blockage</v-chip> byproduct has nowhere to go, so it fills the machine's output and stalls the buildings making it. Blend it into a recipe that consumes it, export it, or sink it.
-    </p>
-    <product :factory="factory" :help-text="helpText" />
+    <product :factory="factory" />
     <v-btn
       color="primary mr-2 mt-n1"
       prepend-icon="fas fa-cube"
@@ -25,15 +16,27 @@
     >
       Add Product
     </v-btn>
-    <power-producer :factory="factory" :help-text="helpText" />
+    <power-producer :factory="factory" />
     <v-btn
-      color="yellow-darken-3 mr-2 mt-n1"
+      class="mr-2 mt-n1"
+      :color="sfColors.powerGeneration.color"
       prepend-icon="fas fa-bolt"
       ripple
       variant="flat"
       @click="addEmptyPowerProducer(factory)"
     >
       Add Power Generator
+    </v-btn>
+    <custom-building :factory="factory" />
+    <v-btn
+      class="mr-2 mt-n1"
+      :color="sfColors.building.color"
+      prepend-icon="fas fa-building"
+      ripple
+      variant="flat"
+      @click="addEmptyCustomBuilding(factory)"
+    >
+      Add Custom Building
     </v-btn>
   </div>
 </template>
@@ -42,12 +45,13 @@
   import { computed } from 'vue'
   import { Factory, FactoryPowerChangeType } from '@/interfaces/planner/FactoryInterface'
   import { addProductToFactory } from '@/utils/factory-management/products'
+  import { sfColors } from '@/utils/colors'
   import { addPowerProducerToFactory } from '@/utils/factory-management/power'
+  import { addCustomBuildingToFactory } from '@/utils/factory-management/custom-buildings'
   import { FactoryStatus, getSectionStatuses, highestSeverity } from '@/utils/factory-management/status'
 
   const props = defineProps<{
     factory: Factory;
-    helpText: boolean;
     statuses?: FactoryStatus[];
   }>()
 
@@ -75,6 +79,10 @@
       recipe: '',
       updated: FactoryPowerChangeType.Power,
     })
+  }
+
+  const addEmptyCustomBuilding = (factory: Factory) => {
+    addCustomBuildingToFactory(factory)
   }
 
   const updateOrder = (list: any[], direction: 'up' | 'down', item: any) => {